GCC warnings

+++ sdk/core/core/test/cmocka/test_az_url_encode.c
@@ -365,13 +365,13 @@ static void test_url_encode_full(void** state)
   (void)state;
 
   uint8_t values256[256] = { 0 };
-  for (int i = 0; i < _az_COUNTOF(values256); ++i)
+  for (size_t i = 0; i < _az_COUNTOF(values256); ++i)
   {
     values256[i] = (uint8_t)i;
   }
 
   uint8_t buf[256 * 3] = { 0 };
-  for (int i = 0; i < _az_COUNTOF(buf); ++i)
+  for (size_t i = 0; i < _az_COUNTOF(buf); ++i)
   {
     buf[i] = '*';
   }
